+++ vec_math.h
@@ -3,28 +3,28 @@
 
 //3차원 벡터
 struct vec3 {
-	double x, y, z;
+	float x, y, z;
 
 	vec3();
 
-	vec3(double x, double y, double z);
+	vec3(float x, float y, float z);
 
 	vec3 operator-() const;
 	vec3 operator+(const vec3& o) const;
 	vec3 operator-(const vec3& o) const;
-	vec3 operator*(const double& k) const;
-	vec3 operator/(const double& k) const;
+	vec3 operator*(const float& k) const;
+	vec3 operator/(const float& k) const;
 
 	vec3& operator+=(const vec3& o);
 	vec3& operator-=(const vec3& o);
 
-	double size() const;
+	float size() const;
 
 	vec3 normalized() const;
 
 	void normalize();
 
-	static double dot(const vec3& a, const vec3& b);
+	static float dot(const vec3& a, const vec3& b);
 
 	static vec3 cross(const vec3& a, const vec3& b);
 
@@ -35,23 +35,23 @@ struct vec3 {
 
 //회전 처리를 위한 사원수
 struct Quat {
-	double w;
+	float w;
 	vec3 v;
 
 	Quat();
-	Quat(double w, vec3 v);
-	Quat(double w, double x, double y, double z);
+	Quat(float w, vec3 v);
+	Quat(float w, float x, float y, float z);
 
-	double sq_norm() const;
+	float sq_norm() const;
 	Quat conj() const;
 
 	Quat operator-() const;
-	Quat operator*(const double& k) const;
-	Quat operator/(const double& k) const;
+	Quat operator*(const float& k) const;
+	Quat operator/(const float& k) const;
 	Quat operator*(const Quat& o) const;
 	Quat operator*(const vec3& o) const;
 
-	static Quat Euler(double x, double y, double z);
+	static Quat Euler(float x, float y, float z);
 	static Quat Euler(vec3 r);
 
 	static vec3 rotate(const vec3& v, const Quat& q);
